Moved mockChrome object into utils folder
